795: Replacing calls to RS to get DebtorAccount with configuration
https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/795
