+++ gradle/profiles/profile-dev.gradle.kts
@@ -51,6 +51,7 @@ val igServer by extra("https://obdemo.$environment.forgerock.financial")
 val userId by extra ("4737f9f9-fa0a-4159-bc61-7da31542e624")
 val userPassword by extra("0penBanking!")
 val username by extra("psu4test")
+val userDebtorAccountIdentification by extra("01233243245676")
 
 // Kid's
 val eidasTestSigningKid by extra("zPeFbX7nJokEVpynzayWvgtFBxo")
