+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/domestic/vrp/consents/api/v3_1_10/CreateDomesticVrpConsents.kt
@@ -216,7 +216,7 @@ class CreateDomesticVrpConsents(val version: OBVersion, val tppResource: CreateT
     }
 
     private fun populateDebtorAccount(consentRequest: OBDomesticVRPConsentRequest) {
-        val debtorAccount = RsUserData().getDebtorAccount()
+        val debtorAccount = PsuData().getDebtorAccount()
         consentRequest.data.initiation.debtorAccount(
             OBCashAccountDebtorWithName()
                 .identification(debtorAccount?.Identification)
